Removed 'id' from Node and simplified the API of Node
The 'id' field was not really necessary in Node as it is only used for
persistence purposes. Each node has its other unique identifier - the
address in memory.

The removing of 'id' allowed for the simplification of the Node API. Now
nodes can be constructed in a more traditional way. Supplying a model
when constructing a node is also now optional.

Test cases for ModelBase and FilePersistence were adjusted to pass for
the new changes. The rest of Envision still needs to be adjusted for
these changes.

Note: the old API only allowed the creation of tree structures in a
top-down fashion. This restriction no longer applies.