Fix two bugs with expression editor handling
- When calling createFinished on an unfinished operator, the underlying
descriptor is sometimes created from scratch thereby preventing the use
of descriptor classes which inherit OperatorDescriptor. This is true for
error descriptors. The fix changes the createFinished method to never
make a copy of the descriptor but just the existing value. In case of
transient descriptors, the unfinished operator is set not to delete its
descriptor when destroyed.

- Fix a bug where finishOperator instructions are merged incorrectly.
Now such instructions are merged only after an AddError is
encountered.