Introduce scene cursors and migrate Text based items to use a TextCursor
This commit extends a Scene to contain cursors. Each scene also has a
main cursor. A cursor is associated with a graphical item and contains
data about the user focal point within that item. An example is the
blinking bar when editing text.

The currently focused item should always be associated with the main
scene cursor, while additional cursors are optional and may for example
come from the interactions of other users.