Fix more bugs and improve global cursor behavior
Fix bugs with not being able to get to various positions within a
layout. Navigating now works as expected almost always.

Introduce a reference point for navigating in different directions. The
system will what was the last cursor position (middle of region) where
the user started navigating using the arrow keys. In case the user keeps
navigating on the same axis, the original reference point is used to
select the next cursor. This greatly improves the precision of the
cursor. This is basically the equivalent to the following when it comes
to text:
AAAAAAA| <--- start here
AAA| <--- press down to get to here
AAAAAA| <--- and once again to get to here
   ^ <--- as opposed to here, which is closer to the previous cursor
position, but further away from the original location where the user
pressed the down arrow.