Add expression editing support for basic arithmetic expressions
This commit introduces the following:
- Add an Adapter Framework similar to functionality in Eclipse. It
allows and instance of a class A to be adapted to an instance of a class
B, without modification of any of them.
- The framework is used to adapt Node classes to a newly introduced
StringComponents class that converts a node to a list of strings -
components.
- The framework is also used to adapt Item classes to a newly introduced
StringProvider class that allows one to get and set the cursor offset of
a linear representation class.
- This functionality is used within the HOOExpression handler to process
key events and edit expressions.
- The cursor is set separately by triggering a new custom event type
SetCursorEvent so that it can be executed after the visualization has
been updated.