Generalize the cursor/region framework for all items
So far regions were only used for Layouts, but they were already generic
enough to work for all items. This commit makes this change and removes
old code that deals with focus setting.

The generic implementation in Item.cpp can now correctly handle all
items and their children whenever there is no need for cursors. When
cursors are needed only the region() method of Item should be redefined
as is currently done in SequentialLayout.