Change the mechanism for switching between control flow and normal view
Currently Control Flow visualization is implemented by changing the
default visualization of Method to VMethodCF and VMethod was never used.
This commit introduces a new visualization which switches between
VMethod and VMethodCF when F2 is clicked.

This allows VMethod and VMethodCF to be changed independently. In
particular it makes it possible to enhance the functionality for VMethod
and to introduce handlers for it.