Improve cursor moving
Introduce a notion of cursor location equivalence. When the use presses
an arrow key the cursor is moved to a new cursor location, that is not
equivalent to the original one.

Remove the old idea of RegionOptions, which was so far exclusively used
for proper cursor placement.

Enhance the style of the SequentialLayout to include flags for cursor
region control during region generation.