Implement a workaround for the incorrect event posting in QT
This bug entry details what seems to be a bug in the Qt event handling
routines:
https://bugreports.qt-project.org/browse/QTBUG-24455?page=com.atlassian.jira.plugin.system.issuetabpanels:comment-tabpanel#issue-tabs

In short, posting new events, in an event handler, sometimes disregards
priority and enqueue the new events after all pending events.

To work around this, this commit implements a mechanism to request
action at the end of an event handler, instead of posting events.