Revert changes made in commit 44552136
It is not trivial how to update the purpose of an item whose parent is a
layout, since it is the layout's parent (and not the layout itself) that
needs to recreate the item.

The old commit implemented a simple hack to go around this issue, which
however disregarded a lot of use cases and caused crashes.

This commit reverts those old changes and introduces limited
functionality for changing the visualization purpose of individual
methods within a class.