Use SymbolMatcher to check for name matches during symbol resolution.
The use of this new class greatly speeds up the mass symbol resolution
performed after loading a project. Instead of creating many QRegExp
objects( which was terribly expensive and not necessary) we create
SymbolMatcher objects. These are just string wrappers in case a fixed
string match is needed (often the case).