Introduce VariableObserver for the JavaDebugger This is reduce code duplication between variable tracking and probing.
The idea is as follows:
An Observer has a certain node where it was triggered (by user
interaction), this is for now also the node with the PlotOverlay
attached. From this node we have a unique observerId. For each observer
we store the observed variables. In the nodesObservedBy_ map we store
each location we want to observe with a certain observerId, when such a
location is hit in the debugger we call the handlerFunc_ of the
observers which liten to this location with the values they request.