Fix adapt_classpath.sh
Previously it wouldn't warn if no directory was found as PLUGIN_PATH was
initialized to empty. This lead to confusing error messages from the
basename command.

This commit properly initializes PLUGIN_PATH to the first entry of the
search paths. Additionally it add another path to search which is used
in newer version of ArchLinux/Eclipse.