Change unified data format to TupleSet instead of Graph
TupleSet is the simplest data structure which fullfills all our
requirements. At the same time it provides a easier to think about
model. Furthermore this new implementation is all value based this
reduces possible memory corruption problems. On the downside the passing
data to script requires a bit more copying.