Add a passthrough query for empty commands
This makes it possible to have a union at the end of a query without
much changing the parsing infrastructure.

It might also make sense to do work only on on part of a split and on
the other just pass through, for example to compare what changed by the
other command.