Fix QueryBuilder to work in any case where left & right is Composite
It can be that left and right are both CompositeQueries but their output
and input count is not matching. For example if left is a list and right
is just an operator, left might have more than 1 output but right only
has 1 input.

By making an abstraction we can properly deal with this situation.