Introduce virtual breakpoint ids
This allows us to add as many (virtual) breakpoints on a node as we
want. With this we can achieve nice things:
-Since a breakpoint has an id we can remove the exact breakpoint while
leaving other breakpoint on the same node intact.
->This allows us to have multiple probes and tracked variables on a
single node and theoretically also multiple user visible breakpoints.

By having a single listener method that can be registered to a
breakpoint id we can simplify our breakpoint handling method to finding
all associated ids and calling the corresponding listener. This also
means we can theoretically detach the track variable and probe
implementation from the debugger.