Fix Undo for editing expressions with Compound Objects
Now, after parsing a textual expression, an Envision expression is
created, which does not directly contain the compound expressions, but
rather placeholders. This is so that the new expression can be created
completely outside of the old tree. Once it is created, it is substituted
for the existing expression tree and is now owned by the tree manager.
The tree manager owning the new expression is a prerequisite to extracting
parts of the old one (the compound expressions) and inserting them in the
new one. This is done as the final step of the modification process and
enables proper undo.