Improve atomic cycle detection when merging to work with multi-cycles
The previous version required that all elements of a detected cycle
only have a single dependency, but that is not the case in many practical
cases. For example if a new node is inserted and old nodes are moved to it
and a deletion depends on several of these moves.