Add a new merge component that removes some conflicting deletes
If a branch **completely** deletes a subtree and the other branch moves it
or relabels it (only if the delete is top-level), then let the move or
relabel happen instead of reporting a conflict. The other branch must move
or relabel the top sub-tree node, but can freely change other nodes as
it wishes. The deleting branch is not allowed to move away nodes, only
delete.