Add an opendir command and corresponding FileSystemEntry node.
Using the opendir command, it's possible to open all files in a directory.

For now only files <100kb in size are supported and they are all opened as
text lines. There's still a lot more work needed to make this feature
more useful.

The idea is to gradually introduce better parsers.