Add the skeleton for a new plug-in CppImportLight
This plug-in will implement a new design for importing C++ files.
Instead of using clang to fully parse and understand an entire project,
the idea is to import file one by one, on the lexical level (similar to
how clang-format works). The resulting OOModel will be much closer to a
text file, compared to the current, highly semantic one. This will make
import/export much simpler, and is a better evolutionary path given the
need to interoperate on the level of text for a while.